requirements.txt
Django==2.2.3
djangorestframework==3.9.4
Pillow==6.0.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
